requirements.txt
tensorflow-gpu==2.10.0
keras==2.10
keras-tcn==3.5.0
keras-flops==0.1.2
xlrd==2.0.1
openpyxl==3.1.2
seaborn==0.12.2
livelossplot==0.5.5
xlsxwriter==3.1.9

numpy~=1.23.5
pandas~=1.3.4
scikit-learn~=1.2.2
matplotlib~=3.6.2
Ipython
scipy~=1.10.0